Update GameStructure document
Changed the Point2D type for the points array in the Block class to just an array of ints. Point2D uses doubles to represent x and y which would be a hassle to cast into ints anytime we work with array indices.
Additionally, decided on 20 rows and 10 columns for the board since that is what it is in standard Tetris and the concrete numbers are useful for writing tests.